% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LISTE_TABELLEN.R
\docType{data}
\name{LISTE_TABELLEN_BUNDESLIGA}
\alias{LISTE_TABELLEN_BUNDESLIGA}
\title{Tabelleninformationen aus Bundesliga und Premier League}
\format{An object of class \code{list} of length 11.}
\usage{
data(LISTE_TABELLEN_BUNDESLIGA); data(LISTE_TABELLEN_PREMIER_LEAGUE)
}
\description{
Die Datensätze LISTE_TABELLEN_BUNDESLIGA bzw. LISTE_TABELLEN_PREMIER_LEAGUE enthalten Tabelleninformationen. Jeder Datensatz
bildet eine verschachtelte Liste. Die erste Ebene ist die Saison. Die zweite Ebene die Spieltage der jeweiligen Saison.
}
\examples{
data(LISTE_TABELLEN_BUNDESLIGA)
data(LISTE_TABELLEN_PREMIER_LEAGUE)
}
\keyword{datasets}
