% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAINFACT_PREMIER_LEAGUE.R
\docType{data}
\name{MAINFACT_PREMIER_LEAGUE}
\alias{MAINFACT_PREMIER_LEAGUE}
\title{Begegnungen Premier League}
\format{An object of class \code{data.frame} with 4180 rows and 117 columns.}
\usage{
data(MAINFACT_PREMIER_LEAGUE)
}
\description{
Der Datensatz enthält alle Begegnungen der englischen Premier League über einen Zeitraum von 11 Jahren - Saison 2005/2006 bis einschließlich Saison 2015/2016. Jeder Begegnung sind
Informationen über das jeweilige Spiel, Tabelleninformationen sowie Wettquoten angefügt. Die Tabelleninformationen jeder Zeile verfügen über einen Lag -
sie geben an, welche Tabellenposition die Mannschaften VOR der Begegnung innehaben.
}
\details{
Die Rohdaten für die Aufbereitung der beiden zentralen Datensätze MAINFACT_BUNDESLIGA und MAINFACT_PREMIER_LEAGUE im Paket entstammen der Webseite \href{http://www.football-data.co.uk/data.php}.
Ein Mapping zwischen Spaltenbezeichnungen in den Rohdaten und den Daten dieses Paktes liegt in der Datei CODEBOOK_QUOTA_DATA.
}
\examples{
data(MAINFACT_PREMIER_LEAGUE)
}
\keyword{datasets}
